Fix bug in Prosebot Editor API

+++ prosebot-editor/api/index.php
@@ -9,12 +9,7 @@ function not_found()
 $uri = parse_url($_SERVER['REQUEST_URI'], PHP_URL_PATH);
 $uri = explode('/', $uri);
 
-if (!isset($uri[2]) || !isset($uri[3])) {
-    not_found();
-    exit();
-}
-
-if ($uri[2] !== "templates") {
+if (!isset($uri[2]) || $uri[2] !== "templates") {
     not_found();
     exit();
 }
@@ -23,10 +18,18 @@ require_once(__DIR__ . '/TemplateController.php');
 $controller = new TemplateController();
 switch ($_SERVER['REQUEST_METHOD']) {
     case 'DELETE': {
+        if (!isset($uri[3])) {
+            not_found();
+            exit();
+        }
         $controller->delete($uri[3]);
         break;
     }
     case 'GET': {
+        if (!isset($uri[3])) {
+            not_found();
+            exit();
+        }
         if (isset($uri[4])) {
             $controller->{$uri[4]}($uri[3]);
         } 
@@ -40,6 +43,10 @@ switch ($_SERVER['REQUEST_METHOD']) {
         break;
     }
     case 'PUT': {
+        if (!isset($uri[3])) {
+            not_found();
+            exit();
+        }
         if (isset($uri[4])) {
             $controller->{"update_" . $uri[4]}($uri[3]);
         } else {
